almost no change

+++ src/calc.rs
@@ -1,10 +1,7 @@
 use std::ops::{Div, Mul};
 /// This function takes a percentage and converts it back to the values.
 
-///  assert_eq!(25,super::percent_to_value(100,25)); 
-/// ```  
-/// # Errors
-/// Incase of division by zero this function will return 'None'
+
 pub fn percent_to_value(total:f64,value:f64)->Option<f64>{
     if total == 0_f64 || value == 0_f64 {
         return None
@@ -13,13 +10,13 @@ pub fn percent_to_value(total:f64,value:f64)->Option<f64>{
     let b = a.mul(100_f64);
     Some(b)
 }
-pub fn percentage(total:f64,value:f64)->Option<u128>{
+pub fn percentage(total:f64,value:f64)->Option<f64>{
     if total == 0_f64 || value == 0_f64 {
         return None
     }
     let a = total.div(100_f64);
     let b = a.mul(value);
-    Some(b.abs() as u128)
+    Some(b.abs())
 }
 //================================================
 //================================================
@@ -70,30 +67,30 @@ fn the_loop(){
 #[cfg(test)]
 #[test]
 fn div_by_zero_perc(){
-    let a:Option<u128> = percentage(0_f64,600_f64);
+    let a:Option<f64> = percentage(0_f64,600_f64);
     assert!(a==None);
 }
 #[cfg(test)]
 #[test]
 fn nominator_zero_perc(){
-    let a:Option<u128> = percentage(60_f64,0_f64);
+    let a:Option<f64> = percentage(60_f64,0_f64);
     assert!(a==None);
 }
 #[cfg(test)]
 #[test]
 fn both_zero_perc(){
-    let a:Option<u128> = percentage(0_f64,0_f64);
+    let a:Option<f64> = percentage(0_f64,0_f64);
     assert!(a==None);
 }
 #[cfg(test)]
 #[test]
 fn the_loop_perc(){
     for i in 1..100 {
-        let a:Option<u128> = percentage(200_f64,i as f64);
+        let a:Option<f64> = percentage(200_f64,i as f64);
         let ans = 200_f64 / 100_f64;
         let ans2 = ans * i as f64;
         match a {
-            Some(x)=>assert!(x == ans2 as u128),
+            Some(x)=>assert!(x == ans2),
             None=>(),
         }
